# xiaoyi/plugin_openapi_v6.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "优活 Agent API",
    "description": "面向独居老人的家庭协同式可信生活智能体。语言层只负责理解；办事、循环事务、情绪陪伴、记忆、健康、用药、位置安全与家庭协同由可审计代码控制；v5新增语音共识、目的绑定策略、可恢复Saga、离线同步、破窗访问与可验证证明；v6新增认知负荷治理、玻璃盒依赖校准、安全预演、受约束语义网关与真实用户实验工具链。",
    "version": "6.0.0"
  },
  "paths": {
...
    "/v5/voice/resolve": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Resolve Voice",
        "operationId": "resolve_voice_v5_voice_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTurnResolution"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/actions/authorize": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Authorize Action",
        "operationId": "authorize_action_v5_actions_authorize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionAuthorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionAuthorization"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/sagas": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Create Saga",
        "operationId": "create_saga_v5_sagas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SagaCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "List Sagas",
        "operationId": "list_sagas_v5_sagas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "elder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Elder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SagaRecord"
                  },
                  "title": "Response List Sagas V5 Sagas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v5/sagas/{saga_id}": {
      "get": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Get Saga",
        "operationId": "get_saga_v5_sagas__saga_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "saga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saga Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v5/sagas/{saga_id}/advance": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Advance Saga",
        "operationId": "advance_saga_v5_sagas__saga_id__advance_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "saga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saga Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SagaAdvanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v5/sync/operations": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Apply Sync",
        "operationId": "apply_sync_v5_sync_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncOperationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncOperationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/sync/conflicts": {
      "get": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "List Sync Conflicts",
        "operationId": "list_sync_conflicts_v5_sync_conflicts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conflict_status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(open|keep_current|accept_incoming)$",
              "default": "open",
              "title": "Conflict Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncConflictRecord"
                  },
                  "title": "Response List Sync Conflicts V5 Sync Conflicts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v5/sync/conflicts/resolve": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Resolve Sync Conflict",
        "operationId": "resolve_sync_conflict_v5_sync_conflicts_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConflictResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncOperationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/break-glass": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Open Break Glass",
        "operationId": "open_break_glass_v5_break_glass_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakGlassRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakGlassRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/break-glass/{elder_id}": {
      "get": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "List Break Glass",
        "operationId": "list_break_glass_v5_break_glass__elder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Elder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BreakGlassRecord"
                  },
                  "title": "Response List Break Glass V5 Break Glass  Elder Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v5/break-glass/{record_id}/close": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Close Break Glass",
        "operationId": "close_break_glass_v5_break_glass__record_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakGlassRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v5/proofs/verify": {
      "post": {
        "tags": [
          "v5 trustworthy elder agent"
        ],
        "summary": "Verify Proof",
        "operationId": "verify_proof_v5_proofs_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofVerifyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoiceTurnRequest": {
        "properties": {
          "elder_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Elder Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/TranscriptCandidate"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Candidates"
          },
          "expected_domain": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Domain"
          },
          "side_effect_possible": {
            "type": "boolean",
            "title": "Side Effect Possible",
            "default": false
          },
          "current_task_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Task Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "elder_id",
          "candidates"
        ],
        "title": "VoiceTurnRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}